refactor: Separate ItemSelectorView and each components
