+++ app/src/main/java/com/mashup/torchlight/ui/customview/itemselectorview/ItemSelectorData.kt
@@ -1,0 +1,10 @@
+package com.mashup.torchlight.ui.customview.itemselectorview
+
+import android.graphics.drawable.Drawable
+
+data class ItemSelectorData(
+    val id: Int,
+    var name: String,
+    var image: Drawable?,
+    var isSelected: Boolean = false
+)
